Ordering System For Restaurants / Java Mini Project
The main goal of this project is to create a software "restaurant management system" that will save time for the clients and the management team of the restaurant. On the other hand, it will help manage business much more effectively than in the past. With this system, customers can order food easily checkout immediately.
Made By Hamdi Guerbej